헤더, GPS 버튼 완성
