+++ src/components/css/header.css
@@ -1,7 +1,7 @@
 .header {
     position: fixed;
     justify-content: center; 
-    padding-top: 20px; 
+    padding-top: 30px; 
     z-index: 999;
   }
